# packages/bc-wallet-openapi/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Credential Showcase API
  version: 0.1.1
servers:
  - url: https://virtserver.swaggerhub.com/SphereonInt/showcase/0.1.0/v1
    description: SwaggerHub API Auto Mocking
  - url: https://bcshowcase-api-dev.nborbit.ca/
...
security:
  - OAuth2: [ ]
  - OIDC: [ ]
paths:
...
  /credentials/definitions:
    get:
      tags:
        - Credential Definitions
      summary: List credential definitions
      operationId: listCredentialDefinitions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialDefinitionsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Credential Definitions
      summary: Create a credential definition
      operationId: createCredentialDefinition
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialDefinitionRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialDefinitionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /credentials/definitions/import:
    post:
      tags:
        - Credential Definitions
      summary: Import a credential definition (for RP only)
      operationId: importCredentialDefinition
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialDefinitionImportRequest'
        required: true
      responses:
        '204':
          description: Imported
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /credentials/definitions/{definitionId}:
    parameters:
      - name: definitionId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Credential Definitions
      summary: Get a specific credential definition
      operationId: getCredentialDefinition
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialDefinitionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Credential Definitions
      summary: Update a credential definition
      operationId: updateCredentialDefinition
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialDefinitionRequest'
        required: true
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialDefinitionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Credential Definitions
      summary: Delete a credential definition
      operationId: deleteCredentialDefinition
      responses:
        '204':
          description: Deleted
...
  /assets:
    get:
      tags:
        - Assets
      summary: List assets
      operationId: listAssets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Assets
      summary: Create an asset
      operationId: createAsset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /assets/{assetId}:
    parameters:
      - name: assetId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Assets
      summary: Get a specific asset
      operationId: getAsset
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Assets
      summary: Update an asset
      operationId: updateAsset
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetRequest'
        required: true
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Assets
      summary: Delete an asset
      operationId: deleteAsset
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /assets/{assetId}/file:
    parameters:
      - name: assetId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Assets
      summary: Retrieve a specific asset as file
      operationId: getAssetAsFile
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: <https://auth.example.com/oauth/authorize>
          tokenUrl: <https://auth.example.com/oauth/token>
          scopes:
            editor: Grants editor access
            admin: Grants admin access
            viewer: Grants viewer access
    OIDC:
      type: openIdConnect
      openIdConnectUrl: <https://auth.example.com/.well-known/openid-configuration>
      description: OpenID Connect security scheme for OAuth2 flows
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the user
          example: 123e4567-e89b-12d3-a456-426614174456
        userName:
          type: string
        issuer:
          type: string
        clientId:
          type: string
        createdAt:
          type: string
          description: Date and time the asset was created
          format: date-time
          example: 2025-03-06T11:30:16.573Z
        updatedAt:
          type: string
          description: Date and time the asset was created
          format: date-time
          example: 2025-03-06T11:30:16.573Z
    Asset:
      type: object
      required:
        - id
        - mediaType
        - content
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier for the asset in the database
          example: 123e4567-e89b-12d3-a456-426614174469
        mediaType:
          type: string
          description: Type of asset (e.g., image, document, icon)
          example: image/jpeg
        content:
          type: string
          description: Binary data of the asset
          example: base64 encoded binary data
        fileName:
          type: string
          description: Name of the asset file
          example: asset.jpg
        description:
          type: string
          description: Description of the asset
          example: A beautiful image of a cat
        createdAt:
          type: string
          description: Date and time the asset was created
          format: date-time
          example: 2025-03-06T11:30:16.573Z
        updatedAt:
          type: string
          description: Date and time the asset was last updated
          format: date-time
          example: 2025-03-06T11:30:16.573Z
    AssetRequest:
      type: object
      required:
        - mediaType
        - content
      properties:
        mediaType:
          type: string
          description: Type of asset (e.g., image, document, icon)
          example: image/jpeg
        content:
          type: string
          description: Binary data of the asset
          example: base64 encoded binary data
        fileName:
          type: string
          description: Name of the asset file
          example: asset.jpg
        description:
          type: string
          description: Description of the asset
          example: A beautiful image of a cat
    AssetsResponse:
      type: object
      required:
        - assets
      properties:
        assets:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
    AssetResponse:
      type: object
      required:
        - asset
      properties:
        asset:
          $ref: '#/components/schemas/Asset'
...
    IdentifierType:
      type: string
      enum:
        - DID
      description: Types of supported identifiers
      example: DID
    Source:
      type: string
      enum:
        - IMPORTED
        - CREATED
      description: Types of supported origins
      example: CREATED
      default: CREATED
...
    CredentialType:
      type: string
      enum:
        - ANONCRED
      description: Types of supported credentials
    CredentialAttributeType:
      type: string
      enum:
        - STRING
        - INTEGER
        - FLOAT
        - BOOLEAN
        - DATE
      description: Types of credential attributes
      example: STRING
    CredentialDefinition:
      type: object
      required:
        - id
        - name
        - version
        - type
        - credentialSchema
        # TODO enable back in SHOWCASE-81
        # - representations
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier for the credential definition
          example: 123e4567-e89b-12d3-a456-426614174123
        name:
          type: string
          description: Name of the credential definition
          example: Credential Definition Name
        credentialSchema:
          $ref: '#/components/schemas/CredentialSchema'
        identifierType:
          $ref: '#/components/schemas/IdentifierType'
        identifier:
          type: string
          description: External identifier of this issuer
          example: did:sov:XUeUZauFLeBNofY3NhaZCB
        version:
          type: string
          description: Version of the credential definition
          example: '1.0'
        type:
          $ref: '#/components/schemas/CredentialType'
        source:
          $ref: '#/components/schemas/Source'
        representations:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/CredentialRepresentation'
              - $ref: '#/components/schemas/OCARepresentation'
        revocation:
          oneOf:
            - $ref: '#/components/schemas/RevocationInfo'
            - $ref: '#/components/schemas/AnonCredRevocation'
        icon:
          $ref: '#/components/schemas/Asset'
        createdAt:
          type: string
          description: Date and time the credential definition was created
          format: date-time
          example: 2025-03-06T11:30:16.573Z
        updatedAt:
          type: string
          description: Date and time the credential definition was last updated
          format: date-time
          example: 2025-03-06T11:30:16.573Z
        approvedBy:
          $ref: '#/components/schemas/User'
        approvedAt:
          type: string
          description: Date and time the credential definition was last approved
          format: date-time
          example: 2025-03-06T11:30:16.573Z
    CredentialDefinitionRequest:
      type: object
      required:
        - name
        - version
        - type
        - credentialSchema
        # TODO enable back in SHOWCASE-81
        # - representations
      properties:
        name:
          type: string
          description: Name of the credential definition
          example: Credential Definition Name
        identifierType:
          $ref: '#/components/schemas/IdentifierType'
        identifier:
          type: string
          description: External identifier of this schema
          example: did:sov:XUeUZauFLeBNofY3NhaZCB
        version:
          type: string
          description: Version of the credential definition
          example: '1.0'
        type:
          $ref: '#/components/schemas/CredentialType'
        source:
          $ref: '#/components/schemas/Source'
        credentialSchema:
          type: string
          description: Reference to the associated credential schema
          example: 456e4567-e89b-12d3-a456-426614174789
        representations:
          type: array
          description: List of representations for this credential definition
          example:
            - type: CREDENTIAL_DEFINITION
          items:
            oneOf:
              - $ref: '#/components/schemas/CredentialRepresentation'
              - $ref: '#/components/schemas/OCARepresentation'
        revocation:
          oneOf:
            - $ref: '#/components/schemas/RevocationInfo'
            - $ref: '#/components/schemas/AnonCredRevocation'
          description: Revocation information for the credential
          example:
            title: Revocation Information
            description: This credential is revocable
        icon:
          type: string
          example: 123e4567-e89b-12d3-a456-426614174469
        approvedBy:
          type: string
          description: Reference to the associated user who approved this definition
          example: 456e4567-e89b-12d3-a456-426614174789
        approvedAt:
          type: string
          description: Date and time the credential definition was last approved
          format: date-time
          example: 2025-03-06T11:30:16.573Z
    CredentialDefinitionsResponse:
      type: object
      properties:
        credentialDefinitions:
          type: array
          description: "List of credential definitions"
          items:
            $ref: '#/components/schemas/CredentialDefinition'
    CredentialDefinitionResponse:
      type: object
      properties:
        credentialDefinition:
          $ref: '#/components/schemas/CredentialDefinition'
    CredentialDefinitionImportRequest:
      type: object
      required:
        - name
        - identifierType
        - identifier
      properties:
        identifierType:
          $ref: '#/components/schemas/IdentifierType'
        identifier:
          type: string
          description: External identifier of this credential definition
          example: did:sov:XUeUZauFLeBNofY3NhaZCB
        name:
          type: string
          description: Name of the credential definition
          example: Credential Definition Name
        version:
          type: string
          description: Version of the credential schema
          example: '1.0'
        icon:
          type: string
          description: Reference to an asset ID for the icon
          example: 123e4567-e89b-12d3-a456-426614174469
        approvedBy:
          type: string
          description: Reference to the user ID who approved this definition
          example: 456e4567-e89b-12d3-a456-426614174789
    CredentialAttribute:
      type: object
      required:
        - id
        - name
        - type
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier for the attribute
          example: 890e4567-e89b-12d3-a456-426614174123
        name:
          type: string
          description: Name of the attribute
          example: name
        value:
          type: string
          description: Default or example value of the attribute
          example: John Doe
        type:
          $ref: '#/components/schemas/CredentialAttributeType'
        createdAt:
          type: string
          description: Date and time the credential attribute was created
          format: date-time
          example: 2025-03-06T11:30:16.573Z
        updatedAt:
          type: string
          description: Date and time the credential attribute was last updated
          format: date-time
          example: 2025-03-06T11:30:16.573Z
...
    CredentialRepresentation:
      type: object
      required:
        - id
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier for the credential representation
          example: 123e4567-e89b-12d3-abcd-426614174456
        createdAt:
          type: string
          description: Date and time the credential representation was created
          format: date-time
          example: 2025-03-06T11:30:16.573Z
        updatedAt:
          type: string
          description: Date and time the credential representation was last updated
          format: date-time
          example: 2025-03-06T11:30:16.573Z
    OCARepresentation:
      allOf:
        - $ref: '#/components/schemas/CredentialRepresentation'
        - type: object
          required:
            - credDefId
            - schemaId
          properties:
            credDefId:
              type: string
              description: Credential definition identifier in the OCA format
              example: 123e4567-e89b-12d3-a456-426614174123
            schemaId:
              type: string
              description: Schema identifier in the OCA format
              example: 123e4567-e89b-12d3-a456-426614174123
            ocaBundleUrl:
              type: string
              description: URL to the OCA bundle for this representation
              example: https://example.com/ocaBundle.json
      description: Overlay Capture Architecture representation of a credential
    RevocationInfo:
      type: object
      required:
        - id
        - title
        - description
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Id of the revocation information
          example: abcd4567-e89b-12d3-a456-426614174123
        title:
          type: string
          description: Title of the revocation information
          example: Revocation Information
        description:
          type: string
          description: Detailed description of the revocation process or requirements
          example: This credential is revocable
        createdAt:
          type: string
          description: Date and time the Revocation info was created
          format: date-time
          example: 2025-03-06T11:30:16.573Z
        updatedAt:
          type: string
          description: Date and time the Revocation info was last updated
          format: date-time
          example: 2025-03-06T11:30:16.573Z
    AnonCredRevocation:
      allOf:
        - $ref: '#/components/schemas/RevocationInfo'
        - type: object
          required:
            - registryId
          properties:
            registryId:
              type: string
              description: Identifier of the revocation registry for AnonCreds
              example: 123e4567-e89b-12d3-a456-426614174123
      description: Anonymous Credentials specific revocation information
    CredentialSchema:
      type: object
      required:
        - id
        - name
        - version
        - attrNames
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier for the credential schema
          example: 123e4567-e89b-12d3-a456-426614174123
        name:
          type: string
          description: Name of the credential schema
          example: Diploma Schema
        version:
          type: string
          description: Version of the credential schema
          example: '1.0'
        identifierType:
          $ref: '#/components/schemas/IdentifierType'
        identifier:
          type: string
          description: External identifier of this schema
          example: did:sov:XUeUZauFLeBNofY3NhaZCB
        source:
          $ref: '#/components/schemas/Source'
        attributes:
          type: array
          items:
            $ref: '#/components/schemas/CredentialAttribute'
        createdAt:
          type: string
          description: Date and time the credential attribute was created
          format: date-time
          example: 2025-03-06T11:30:16.573Z
        updatedAt:
          type: string
          description: Date and time the credential attribute was last updated
          format: date-time
          example: 2025-03-06T11:30:16.573Z
...
  responses:
    'BadRequest':
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Invalid request data'
    'Unauthorized':
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Invalid authentication credentials'
    'Forbidden':
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'Insufficient permissions to access resource'
...
    'InternalServerError':
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: 'An unexpected error occurred'